app/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cloudinary==1.41.0
cryptography==44.0.0
Django==5.1.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
